pyproject.toml
[tool.poetry]
name = "polymag_task_manager"
version = "0.1.0"
description = ""
authors = ["Maxim P <maxxxlim@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.1.1"
rollbar = "^0.15.0"
gunicorn = "^20.0.4"
django-bootstrap4 = "^2.2.0"
python-dotenv = "^0.15.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.0.1"
isort = "^5.5.1"
black = "^20.8b1"
coverage = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
